+++ pkg/aws/v1alpha3/cluster/validate_cluster_test.go
@@ -519,3 +519,105 @@ func TestValidClusterStatus(t *testing.T) {
 		})
 	}
 }
+
+func Test_CiliumReleaseIsValid(t *testing.T) {
+	testCases := []struct {
+		name           string
+		currentRelease string
+		targetRelease  string
+		annotation     map[string]string
+
+		valid bool
+	}{
+		{
+			name:           "case 0: no cilium",
+			currentRelease: "17.4.0",
+			targetRelease:  "17.4.1",
+			valid:          true,
+		},
+		{
+			name:           "case 1: upgrade to cilium with pod cidr annotation",
+			currentRelease: "17.4.1",
+			targetRelease:  "18.0.0",
+			annotation:     map[string]string{annotation.CiliumPodCidr: "10.0.0.0/8"},
+
+			valid: true,
+		},
+		{
+			name:           "case 2: upgrade to cilium without pod cidr annotation",
+			currentRelease: "17.4.1",
+			targetRelease:  "18.0.0",
+
+			valid: false,
+		},
+		{
+			name:           "case 3: upgrade to cilium without pod cidr annotation",
+			currentRelease: "18.0.0",
+			targetRelease:  "18.1.0",
+
+			valid: true,
+		},
+	}
+
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			v := &Validator{
+				k8sClient: unittest.FakeK8sClient(),
+				logger:    microloggertest.New(),
+			}
+			cluster := unittest.DefaultCluster()
+			cluster.SetLabels(map[string]string{
+				label.Cluster:        unittest.DefaultClusterID,
+				label.ReleaseVersion: tc.targetRelease,
+			})
+
+			oldCluster := unittest.DefaultCluster()
+			oldCluster.SetLabels(map[string]string{
+				label.Cluster:        unittest.DefaultClusterID,
+				label.ReleaseVersion: tc.currentRelease,
+			})
+
+			// create releases for testing
+			releases := []unittest.ReleaseData{
+				{
+					Name: "v17.4.0",
+				},
+				{
+					Name: "v17.4.1",
+				},
+				{
+					Name: "v18.0.0",
+				},
+				{
+					Name: "v18.1.0",
+				},
+			}
+			for _, r := range releases {
+				release := unittest.DefaultRelease()
+				release.SetName(r.Name)
+				err := v.k8sClient.CtrlClient().Create(context.Background(), &release)
+				if err != nil {
+					t.Fatal(err)
+				}
+			}
+			awsCluster := unittest.DefaultAWSCluster()
+			awsCluster.SetAnnotations(tc.annotation)
+			awsCluster.SetLabels(map[string]string{
+				label.ReleaseVersion: "v" + tc.currentRelease,
+				label.Cluster:        unittest.DefaultClusterID,
+			})
+			err := v.k8sClient.CtrlClient().Create(context.Background(), awsCluster)
+			if err != nil {
+				t.Fatal(err)
+			}
+			// check if the result is as expected
+			err = v.Cilium(oldCluster, cluster)
+			if tc.valid && err != nil {
+				t.Fatalf("unexpected error %v", err)
+			}
+			if !tc.valid && err == nil {
+				t.Fatalf("expected error but returned %v", err)
+			}
+		})
+	}
+}