+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validate `cilium.giantswarm.io/pod-cidr` annotation is present and valid while upgrading from v17 to v18.
+
 ## [4.1.0] - 2022-07-22
 
 ## [4.0.3] - 2022-06-15